Updated RatableBehavior::calculateRating()
Fixed issue where unnecessary Model::save() is called when the $saveToField does not exist in the model schema.

Not checking the model schema for $saveToField would cause two Model::exists() calls - one in Model::save() and one in Model::invalidFields() - and finally an UPDATE statement which would merely updated the `modified` field, rendering this method useless.